+++ apps/example-bot/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @signald-node/example-bot
+# @signald/example-bot
 
 ## 0.0.1
 
@@ -6,5 +6,5 @@
 
 - 5e075a9: add reply recipient util function
 - Updated dependencies [5e075a9]
-  - @signald-node/client@0.0.1
-  - @signald-node/bot@0.0.1
+  - @signald/client@0.0.1
+  - @signald/bot@0.0.1